requirements.txt
azure-common==1.1.25
azure-functions==1.2.1
psycopg2-binary==2.8.5
python-http-client==3.3.7
sendgrid==6.3.0
SQLAlchemy==1.3.16